Updated homework
I am trying to pass the argurments through each function to the object constructor at the end, but userPassword and userEmail are not available outside of their functions. I tried defining them outside the functions and setting them when the function is called, but it's like it skips the prompts. Not working yet.